+++ app/models/insight.rb
@@ -4,7 +4,8 @@ class Insight < ApplicationRecord
   DEFAULT_ATTRIBUTES = %i[comments_count reviews_count average_review_seconds open_pull_requests_count].freeze
 
   TIME_ATTRIBUTES = %i[average_review_seconds average_merge_seconds].freeze
-  DECIMAL_ATTRIBUTES = %i[average_open_pr_comments review_involving].freeze
+  DECIMAL_ATTRIBUTES = %i[average_open_pr_comments].freeze
+  PERCENTILE_ATTRIBUTES = %i[review_involving].freeze
 
   REVERSE_ORDER_ATTRIBUTES = %i[average_review_seconds average_merge_seconds average_open_pr_comments].freeze
   SHORT_ATTRIBUTE_NAMES = {
@@ -17,7 +18,9 @@ class Insight < ApplicationRecord
     average_merge_seconds: 'Average merge time',
     average_open_pr_comments: 'Average received comments'
   }.freeze
+
   FETCH_DAYS_PERIOD = 30
+  DOUBLE_FETCH_DAYS_PERIOD = 60
 
   belongs_to :insightable, polymorphic: true
   belongs_to :entity