+++ app/views/components/views/shared/insights_component.rb
@@ -6,9 +6,36 @@ module Views
       def initialize(insightable:)
         @insightable = insightable
         @access_token = insightable.access_token
-        @insights = @insightable.sorted_insights
+        @insights = insightable.sorted_insights
+        @premium = insightable.premium?
+        @insight_ratio = @premium && insightable.configuration.insight_ratio
 
         super()
+      end
+
+      def insight_fields
+        @insight_fields ||=
+          if @premium && @insightable.configuration.insight_fields.present?
+            @insightable.selected_insight_fields
+          else
+            Insight::DEFAULT_ATTRIBUTES
+          end
+      end
+
+      def render_insight_field(insight, attribute)
+        result = convert_insight_field(insight, attribute.to_sym).to_s
+        result += insight_ratio_value(insight, attribute) if @insight_ratio && result != '-'
+        result.html_safe
+      end
+
+      private
+
+      def convert_insight_field(insight, insight_field)
+        return convert_seconds(insight[insight_field]) if Insight::TIME_ATTRIBUTES.include?(insight_field)
+        return insight[insight_field].to_f if Insight::DECIMAL_ATTRIBUTES.include?(insight_field)
+        return "#{insight[insight_field].to_i}%" if Insight::PERCENTILE_ATTRIBUTES.include?(insight_field)
+
+        insight[insight_field].to_i
       end
 
       def convert_seconds(value)
@@ -17,37 +44,37 @@ module Views
 
       # rubocop: disable Layout/LineLength, Rails/OutputSafety
       def insight_ratio_value(insight, attribute)
-        ratio_field_name = "#{attribute}_ratio"
-        ratio_value = insight[ratio_field_name].to_i
-        # for time attributes less value is better
-        ratio_value *= -1 if Insight::TIME_ATTRIBUTES.include?(attribute)
+        return '' if insight[attribute].nil?
 
-        "(<span class='#{ratio_value.negative? ? 'negative' : 'positive'}'>#{ratio_value.positive? ? '+' : ''}#{ratio_value}%</span>)".html_safe
+        ratio_value = insight["#{attribute}_ratio"].to_i
+        time_attribute = Insight::TIME_ATTRIBUTES.include?(attribute.to_sym)
+        change_type = @insightable.configuration.insight_ratio_type == 'change'
+
+        value = multiple_value(ratio_value, time_attribute, change_type)
+        value_for_rendering = time_attribute && change_type ? convert_seconds(value.abs) : value.abs
+
+        " (<span class='#{span_class(ratio_value, time_attribute)}'>#{value_sign(value)}#{value_for_rendering}#{change_type ? '' : '%'}</span>)"
       end
       # rubocop: enable Layout/LineLength, Rails/OutputSafety
 
-      def insight_fields
-        @insight_fields ||=
-          if premium && @insightable.configuration.insight_fields.present?
-            @insightable.selected_insight_fields
-          else
-            Insight::DEFAULT_ATTRIBUTES
-          end
+      def span_class(ratio_value, time_attribute)
+        return 'negative' if ratio_value.negative? && !time_attribute
+        return 'negative' if !ratio_value.negative? && time_attribute
+
+        'positive'
       end
 
-      def convert_insight_field(insight, insight_field)
-        return convert_seconds(insight[insight_field]) if Insight::TIME_ATTRIBUTES.include?(insight_field)
-        return insight[insight_field].to_f if Insight::DECIMAL_ATTRIBUTES.include?(insight_field)
+      def multiple_value(ratio_value, time_attribute, change_type)
+        return ratio_value * -1 if time_attribute && !change_type
 
-        insight[insight_field].to_i
+        ratio_value
       end
 
-      def premium
-        @insightable.premium?
-      end
+      def value_sign(value)
+        return '+' if value.positive?
+        return '-' if value.negative?
 
-      def insight_ratio
-        @insight_ratio ||= premium && @insightable.configuration.insight_ratio
+        ''
       end
     end
   end